Everything is still broken, but I realize what I was doing wrong now
