+++ src/parsers/container/ts/packet.js
@@ -1,8 +1,8 @@
 class Packet {
   constructor(header, bitReader) {
     this.header = header
-    this.data   = bitReader.data
-    this.length = this.data.length
+    this.data   = bitReader.data.slice(4)
+    this.length = this.header.length + this.data.length
   }
 }
