+++ src/parsers/container/ts/parser.js
@@ -21,23 +21,19 @@ class TransportStreamParser {
       const pat   = new PAT(header, bitReader)
       this.pmtPID  = pat.pmtID
       return pat
-    }
-
-    // Header matches the pmtPID we got from the PAT
-    if (header.PID === this.pmtPID) {
+    } else if (header.PID === this.pmtPID) {
       const pmt       = new PMT(header, bitReader)
       this.tracks     = pmt.tracks
       this.trackPIDs  = pmt.tracks.map(t => t.elementaryPID)
       return pmt
-    }
-
-    // Check if the pid matches any of the pids from the tracks we parsed from the PMT
-    if (this.trackPIDs.includes(header.PID)) {
+    } else if (this.trackPIDs.includes(header.PID)) {
       const track       = this.tracks.filter(t => t.elementaryPID === header.PID)[0]
       const mediaPacket = new MediaPacket(header, bitReader, track.streamType)
       return mediaPacket
+    } else {
+      /// This should not have happened
+      console.log('!!! ERROR PARSING PROGRAM PACKETS !!!');
     }
-
   }
 }
